+++ src/HttpHeaderInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace SmartonDev\HttpCache;
-
-interface HttpHeaderInterface
-{
-    public function toHeaders(): array;
-}